Small Re-org, works with guild numbers now. Need to add checker to see if guild actually exists.
